skore/requirements.txt
Flask
Flask-Cors
PyYAML
Requests
